move the `BankLines_Processor` to a separate module `banklines_processor`
